Implement booking with inventory and overbooking rules
1. Add business rules:
quantity > 0
ticket existence and schedule consistency
quantity <= seat inventory
prevent overbooking based on aircraft capacity

2. Restore inventory and cancel booking with status transition
3. Add transactional handling via UnitOfWork

+++ AirportManagement/AirportTool.Application/DTOs/Bookings/CreateBookingDto.cs
@@ -3,7 +3,7 @@
     public class CreateBookingDto
     {
         public int FlightScheduleId { get; set; }
-        public int TicketId { get; set; }
+        public long TicketId { get; set; }
         public string PassengerFullName { get; set; } = string.Empty;
         public string PassengerEmail { get; set; } = string.Empty;
         public int Quantity { get; set; }
